Arduino 1.0

+++ LCDi2c4bit.h
@@ -60,7 +60,7 @@ public:
   void print( char value[] );
   void print( float );
   void print( int );
-  void printL( char value[], int);
+  void printL( char value[], uint8_t);
   void clear();
   void backLight( uint8_t );
   void SendToLCD( uint8_t );
